Styling for About page, Dashboard, Footer, and Home.  Modified Database to include course description.
Reformatted the About page for a grid layout and added some different border and heading styling. Added some spacing in between the courses in the dashboard and added a section for grades which will be implemented in the future. Modified the Footer for the back to top arrow, making it only pop up when the user has scrolled or is close to the bottom of the page. Added course descriptions back into the database so that the descriptions can be displayed on the courses page for our available courses.