Implement Assignments Functionality and Styling Changes
Note: Update your database to properly test the assignments functionality (Assignments table and Mock Data)

Implemented the frontend design for assignments and graded assignments. This needs to be connected to the grades table of our database, once it is created. Adjusted the styling for the Chat, CoursePage, Courses, Dashboard, Login/Signup, and Profile page.